moved book-list to auth screen
