+++ wolox-books/src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import { SignUpComponent } from './screens/unauth/screens/sign-up/sign-up.compon
 import { LogInComponent } from './screens/unauth/screens/log-in/log-in.component'
 import { AuthComponent } from './screens/auth/auth.component'
 import { UnauthComponent } from './screens/unauth/unauth.component';
-import { BookListComponent } from './screens/book-list/book-list.component';
+import { BookListComponent } from './screens/auth/screens/book-list/book-list.component';
 import { AuthGuard } from './auth.guard';
 import { UnauthGuard } from './unauth.guard';
